bmi calculator
(Python) BMI calculator - Final project in my introduction to coding where I learned Python, and GUIs, along with a little advanced object oriented programming. This was one of the first projects that I was able to make on my own, to where the lines were readable, where I understood what actually was happening. This calculator is is actually a bmi which indicates whether or not you are obese or not.